venturepilot-ai/requirements.txt
fastapi>=0.104.0
uvicorn>=0.24.0
pydantic>=2.5.0
mistralai>=1.0.0
python-dotenv>=1.0.0
streamlit>=1.29.0
requests>=2.31.0
numpy>=1.24.0
